pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-wardround-env"
version = "0.1.0"
description = "WardRound-Env: multi-agent hospital ward-round simulator for OpenEnv"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.3",
  "fastapi>=0.115.0",
  "uvicorn>=0.24.0",
  "pydantic>=2.7.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0",
]

[project.scripts]
server = "server.app:main"
run-inference = "inference:main"

[tool.setuptools]
include-package-data = true
packages = ["server"]
